refactor: Move pyproject exceptions to their own submodule
